requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
dj-database-url==2.3.0
django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mailersend==0.5.8
model-bakery==1.20.4
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
protobuf==6.30.2
psycopg2-binary==2.9.10
pyjwt==2.9.0
python-dotenv==1.1.0
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
temporalio==1.10.0
types-protobuf==5.29.1.20250403
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
